refactor: organizing the file flow
