+++ src/contexts/Auth/constants/index.js
@@ -1,0 +1,1 @@
+export { authContextModel } from "./models";
