+++ src/contexts/Auth/constants/models.js
@@ -1,0 +1,4 @@
+export const authContextModel = {
+  profileType: false,
+  setProfileType: () => {},
+};
