+++ src/contexts/Auth/hooks/useAuth.js
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "../contexts/AuthContext";
+import { AuthContext } from "..";
 
 export const useAuth = () => {
   return { ...useContext(AuthContext) };
