+++ src/pages/Register/forms/RegisterWithPhone.jsx
@@ -1,5 +1,5 @@
 import { Form, Input, Button } from "../../../components";
-import { useAuth } from "../../../hooks/useAuth";
+import { useAuth } from "../../../contexts";
 import { validationRegisterWithPhone } from "../validations";
 import * as S from "./RegisterStyled";
